fix: avoid storing aarch64 wasm on mainnet
Aarch64 wasm binary files may be stored on LocalTerra or testnet, however, attempting to store them on mainnet will lead to an error for the user.  If the user attempts to store an aarch64 wasm on mainnet, terrain will now automatically rebuild and reoptimize the contract as a default wasm file for storage.