fix: avoid storing ARM64 wasm on mainnet (#125)
ARM64 wasm binary files may be stored on LocalTerra or testnet, however, attempting to store them on mainnet will lead to an error for the user.  If the user attempts to store an ARM64 wasm on mainnet, terrain will now automatically reoptimize the contract as a default wasm file for storage.