feat: run terrain:contract commands anywhere (#131)
* feat: terrain build anywhere

* feat: run optimize anywhere

* feat: run store anywhere

* feat: run instantiate everywhere

* fix: Specify execute path in warning

Adding execute path to command execute warning.  This is to alert the user of the execute path, just in case this file has been accidentally removed, renamed, or deleted.

* fix: instantiate not deployed alert

Alert user if they run `terrain contract:instantiate` and the object in the `refs.terrain.json` file does not contain the specified network as a key or if the object which is the value of the specified network key does not contain the contract name as a key.  In this case, the contract has not been deployed to the specified network.