style: TerrainCLI (#135)
* style: TerrainCLI

Updating TerrainCLI styling to include stylized, boxed terminal messages.  TerrainCLI logs will now require the user to provide a message as well as a title for presentation.  Furthermore, the user can surround important information with double quotes so that it may be highlighted in a different color.  Example:

`TerrainCLI.error('Contract "${args.contract}" has not been deployed to the "${args.network}" network', 'Contract Not Deployed');`

* fix: Terrain warning at end of code execution